LayoutTests/fast/css/overscroll-behavior-validate-if-enable.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Test overscrollBehavior should be validated if overscrollBehaviorEnabled is enabled");

shouldBeTrue("'overscrollBehavior' in document.documentElement.style");
shouldBeTrue("'overscroll-behavior' in getComputedStyle(document.documentElement)");
</script>
</body>
</html>
