Safari-611.1.10/LayoutTests/fast/css/overscroll-behavior-validate-if-enable-expected.txt

Test overscrollBehavior should be validated if overscrollBehaviorEnabled is enabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'overscrollBehavior' in document.documentElement.style is true
PASS 'overscroll-behavior' in getComputedStyle(document.documentElement) is true
PASS successfullyParsed is true

TEST COMPLETE
